+++ quelware-client/src/quelware_client/client/_standalone_grpc.py
@@ -9,6 +9,8 @@ from quelware_client.infra.diagnostics_agent_grpc import DiagnosticsAgentGrpc
 from quelware_client.infra.instrument_agent_grpc import InstrumentAgentGrpc
 from quelware_client.infra.pat_provider_file import pat_provider_from_config
 from quelware_client.infra.resource_agent_grpc import ResourceAgentGrpc
+from quelware_client.infra.trigger_agent_grpc import TriggerAgentGrpc
+from quelware_client.infra.worker_agent_grpc import WorkerAgentGrpc
 from quelware_client.testing.session_agent_mock import SessionAgentMock
 from quelware_client.testing.system_configuration_agent_mock import (
     SystemConfigurationAgentMock,
@@ -59,9 +61,13 @@ def create_standalone_client(
             channel, metadata={"x-unit-label": str(ul), "x-pat": _pat}
         )
 
+    def worker_agent_factory(ul: UnitLabel):
+        return WorkerAgentGrpc(channel, metadata={"x-unit-label": str(ul)})
+
     agent_container = AgentContainer()
     agent_container.session = session_agent
     agent_container.system_configuration = conf_agent
+    agent_container.trigger = TriggerAgentGrpc(channel)
 
     logger.warning("NOTE: Standalone client is for testing purposes.")
 
@@ -70,6 +76,7 @@ def create_standalone_client(
         resource_agent_factory=resource_agent_factory,
         instrument_agent_factory=instrument_agent_factory,
         diagnostics_agent_factory=diagnostics_agent_factory,
+        worker_agent_factory=worker_agent_factory,
         close_handlers=[channel.close],
         skip_lock_check=skip_lock_check,
     )