+++ quelware-client/src/quelware_client/core/interfaces/instrument_agent/__init__.py
@@ -1,7 +1,6 @@
 from collections.abc import Collection, Sequence
 from typing import Protocol
 
-from quelware_core.entities.clock import CurrentCount, ReferenceCount
 from quelware_core.entities.directives import Directive
 from quelware_core.entities.instrument import InstrumentStatus
 from quelware_core.entities.resource import ResourceId
@@ -34,10 +33,6 @@ class InstrumentAgent(Protocol):
         token: SessionToken,
         resource_ids: Collection[ResourceId],
     ) -> bool: ...
-
-    async def get_clock_snapshot(
-        self,
-    ) -> tuple[CurrentCount, ReferenceCount]: ...
 
     async def schedule_trigger(
         self,