+++ quelware-client/src/quelware_client/core/interfaces/worker_agent/__init__.py
@@ -1,0 +1,7 @@
+from typing import Protocol
+
+from quelware_core.entities.clock import CurrentCount, ReferenceCount
+
+
+class WorkerAgent(Protocol):
+    async def get_clock_snapshot(self) -> tuple[CurrentCount, ReferenceCount]: ...