+++ quelware-client/src/quelware_client/testing/mock_client.py
@@ -8,6 +8,8 @@ from quelware_client.testing.session_agent_mock import SessionAgentMock
 from quelware_client.testing.system_configuration_agent_mock import (
     SystemConfigurationAgentMock,
 )
+from quelware_client.testing.trigger_agent_mock import TriggerAgentMock
+from quelware_client.testing.worker_agent_mock import WorkerAgentMock
 
 
 def create_mock_quelware_client(
@@ -19,6 +21,7 @@ def create_mock_quelware_client(
     agent_container.system_configuration = SystemConfigurationAgentMock(unit_to_status)
 
     agent_container.session = SessionAgentMock()
+    agent_container.trigger = TriggerAgentMock()
 
     def resource_agent_factory(ul: UnitLabel) -> ResourceAgentMock:
         rinfos = unit_to_resources.get(ul, [])
@@ -27,8 +30,12 @@ def create_mock_quelware_client(
     def instrument_agent_factory(ul: UnitLabel) -> InstrumentAgentMock:
         return InstrumentAgentMock()
 
+    def worker_agent_factory(ul: UnitLabel) -> WorkerAgentMock:
+        return WorkerAgentMock()
+
     return QuelwareClient(
         agent=agent_container,
         resource_agent_factory=resource_agent_factory,
         instrument_agent_factory=instrument_agent_factory,
+        worker_agent_factory=worker_agent_factory,
     )